tq-30: fix
