+++ src/validation/validationResult.ts
@@ -1,0 +1,4 @@
+export interface ValidationResult {
+  isValid: boolean
+  error?: string
+}
